Fix a layout issue with scaffolded forms
Scaffolded form input fields have their width set to 100%, but since
their parent has a padding and its box-sizing is set to `inherit`, the
field gets cut off.

This commit adds an extra class from Tailwind CSS and applies it to the
form container to ensure that the entire inputs are visible.