Add/change vars in templates
For the imports/routes I had to add/change the vars in the templates respectively.

Getting imports to work required me to add one--pascalScaffoldPath.
And for routes, I had to make that var in the scaffold.js file,
So I had to add new route vars--e.g. routes.${newRouteName}(), etc.