Rename default datasource
When looking in the documentation for experimental features (For example
using DB native types) you need to add a previewFeatures flag to the
Generate block, and then in the docs they use `@db.*` everywhere.

So as an example if you'd want to use the Postgres native Date field you
get something like this:

`createdAt DateTime @default(now()) @db.Date`

Since we use `DS`as a default datasource it can trip users up a bit
since `@DS.Date` is not a valid Prisma line (Somehow they don't like the
uppercase datasource).

So if we would already rename it to `datasource db {}` we would make it
easier for users in the future