Return signup Output
When you signup via Supabase you get all the meta data back in the
return of the `signup` call. We now eat all that output instead of
giving it back to the user.

I noticed that we already do something similar to Sign In, so I
basically took that concept and applied that for signUp as well.