Allows users to create a custom storybook config
This allows a user to create a custom Storybook preview file in `web/config/storybook.preview.js` and have it imported into the framework's version.

I haven't added any documentation yet for two reasons. First, I want some confirmation that this approach is reasonable. But also, this actually took me a lot longer than it should have because I didn't realize that we're still on Storybook 5.3 and had to use the old `addDecorator` syntax. I suspect this will trip up other users as well, so perhaps it would be worthwhile to update to the latest Storybook before publicizing this feature? Anyone who tries to reference the Storybook docs for the content of this file will likely also be confused.