Static typing for cells
This PR is the first step towards implementing https://github.com/redwoodjs/redwood/issues/2205. It allows users to declare the prop types their cells expect.

It also renames `withCell` to `createCell`, for the reasons I outlined in https://github.com/redwoodjs/redwood/issues/2205#issuecomment-812964440. `createCell` feels like a more parsimonious name to me but I can back that change out if others disagree.

It appears that the babel transform for wrapping cells already checks for a default export and bails out if one exists, so this shouldn't require any changes there.

The major remaining task is updating the cell generator to add a `export default createCell` line at the end of each cell in Typescript. I'd appreciate more input on whether this is a change we're fine making on both the TS and JS sides, or whether we'll need two templates. Do we have multiple templates anywhere else?