Fix issue where Clerk client is sometimes unset
Discovered this while using Clerk + Redwood in production.
Because Clerk is set up during the render cycle by the provider component, in production on first load this value can be captured as `null` by the App.tsx closure. On the client side, this can be fixed by moving Auth Provider to a lower point in the render stack, but the auth setup generator doesn't have support for this. Instead, it can also be fixed by having the `@redwoodjs/auth` package fall back to `window.Clerk` when the client is not defined.