test(e2e): add stderr check
In local testing, `yarn rw storybook --build` returned with a `0` exit code, however there was `stderr`.

This results in no Storybook artifacts being built, which in turn causes subsequent portions of the Cypress test to fail due to lack of the Storybook assets it is expecting to check the existence of.

This newly add `stderr` check should take care of this case and avoid any future false negatives when calling the storybook build in this way.

---

*Additional flavor*

Noticed this while looking into a failure of [another PR I was working on](https://github.com/redwoodjs/redwood/actions/runs/1299581375).

See also: https://discord.com/channels/679514959968993311/716252919875240007/894760524095578205