allow RW auth to use captialised authorization header
Deploying to AWS (with serverless) seems to transform the authorization header to be
captialised "Authorization".

This causes the auth to fail.

I found this from my own experience deploying this test project
https://github.com/Irev-Dev/redwood-serverless-cors-demo

Besides my own experience I tried to find docs confirming this, I found
https://docs.aws.amazon.com/apigateway/latest/developerguide/api-gateway-known-issues.html
Which says:
"Header names and query parameters are processed in a case-sensitive way."

and lists authorization with a capital "A", it doesn't explicitly say
the header is transformed to uppercase, but I think it's implied.

Besides AWS I think it somewhat likely that other deploy targets in the
future might have the same quirk so this only seems to make the auth
provider more robust.