Update CONTRIBUTING.md (#4165)
* Update CONTRIBUTING.md

Two quality of life improvements based on my recent efforts to get set up as a contributor. Working through the doc as a new contributor, I was confused when I got to this section. While reading through the four options it almost seemed like they blended together, and it was not clear that if I chose 1-3 that I should jump to the next section on integrating rwfw. This grouping change attempts to make it more clear that there are four separate options to choose from, that the functional test project may be the best option, and that you can skip to the next section if you choose not to use the functional test project. I do not think that I added any verbiage except updating the headers and TOC appropriately.

Also, I added <cli command> into the `yarn dev <cli command> --cwd <project directory>` since it took me some trial and error to figure out how to use that correctly.

* Apply suggestions from code review

Co-authored-by: David Price <thedavid@thedavidprice.com>