docs: Rename "Serverless.com" menu
(discussed in Slack)

"Serverless.com" is confusing because the company (serverless.com) provides multiple tools, like Serverless Console, Serverless Cloud, etc. which are not involved here.

Instead, the deployment tool being used is "Serverless Framework". So I thought this might be more clear, and more precise, to name the menu that way.