"guides" should link to "Reference" section
Currently, it links to the documentation itself, to the "Introduction" section, which is the home page.
I presume that if the reader arrived here, they probably have already went through it and, hopefully, the whole tutorial.
It would make more sense to provide them with a learn more section such as the "Reference" section as long as "…for more info on Redwood's internals." is referred to that section (pun intended, I couldn't help myself…). 

One final note: the URL for the "Reference" page should probably be `https://redwoodjs.com/docs/reference` for SEO purposes.