configure args type for field resolvers in service
When services are generated for models that have relations, Redwood
generates field resolvers for those relations. This is very helpful, but
defaults to showing alerts in a developer's editor for those lines of
code. These alerts can be avoided by marking the type as `unknown`,
since at the moment that's what it is, and allow users to add type
definitions to the arguments if any arguments are added to the field in
GraphQL.