Make functions opaque, clean up tests
AST nodes can now be marked as "opaque" in the merge strategy, which indicates that the children of the opaque node should not be merged, even if such a child has a merge strategy defined. E.g., we don't want to merge array expressions inside functions, even though we want to merge arrays in general.