Updated testing.md to reflect tutorial flow (#5340)
When working through the tutorial I spotted the failing test and thought I was doing something wrong. I worked through fixing the failure on my own in a slightly different way than outlined in the next section, but if I had seen the tutorial presented this way I would have felt more confident and less confused with what I was doing. This appeared to be a simple fix to the docs, so I thought I would present it.