RFC on using API specs to drive scaffolding
This is a Request for Comment on using API specs to drive scaffolding 

If you are having your front-end team work with an existing API (could be internal brown-field project or external vendor APIs) or if you have separate FE and BE teams working on a project at different paces. In both of these use cases, you want to drive the scaffolding off an API spec rather than a database schema.