chore(CRWA): try publishing crwa with yarn.lock
_I will revert this commit after a canary is published_

this could speed up CRWA and may fix the esbuild problem. to find out I
want to publish a canary. the reason I didnt' do it for v5 was just that
I wasn't sure it was goign to work, and didn't want to introduce another
variable into an important release.

this file was generated by going into the create-redwood-app/template
directory, making an empty yarn.lock file, and running yarn install there.