feat(cli): Experimental support for plugins (#8316)
* Initial plugin loader

* Update project-config test snapshot

* Refactor to attempt to improve readability.

Made CLI TOML options nested under experimental. Log errors for cache IO related tasks. Add checks to ensure package and version are listed in the toml objects. Print a warning if we did't load any plugins when running a command.

* Another refactor and a slight update to the logic.

We should only load the one plugin that is known to match the command via the cache. Load all plugins if we need yargs to show the help.

* chore: correct spelling

---------

Co-authored-by: Dominic Saadi <dominiceliassaadi@gmail.com>