Rework OTel exporting.
Switched to use a custom exporter which writes spans to a file. Then have a background job that fires when OTel shuts down to read those saved spans and send them to the collector.
This means we have one background job that runs to both compute the telemetry resources and send the data to the collector.