chore(cli): Add OpenTelemetry (#8653)
* Initial OTel kernel for CLI

* Add helper functions for telemetry

* Add telemetry for info command

* Capture and record top level errors

* WIP compute of telemetry details in the background

* wip changes to the background compute

* Instrument CLI commands

I'm not in love with this code. Seems difficult to ensure key consisitency but I'm working around the existing code and yargs.

* Rework OTel exporting.

Switched to use a custom exporter which writes spans to a file. Then have a background job that fires when OTel shuts down to read those saved spans and send them to the collector.
This means we have one background job that runs to both compute the telemetry resources and send the data to the collector.

* Re-enable existing telemetry and remove error messaging

* Revert change to locking.js

* Refactor the new telemetry setup

Move some files around and rename some. Adds in the missing telemetry fields to the resource. Rewrites the spans to disk with the added resource information for verbose visibility.

* Remove stray console logs and remove events from spans.

* Produce top level error output and record an error ref. code

We will likely have to iterate on how the output looks.

* fix experiments and webBundler resource values

* Tidy telemetry/index.js

* Record top level errors with legacy telemetry also

* Use error exit code if it exists

I wouldn't have written this myself but I seen it used in the code so I may as well keep using it.

* Started to remove the handler level try/catch.

Stopped because of some many edge or special cases where the CLI immediately exists with non-zero exit code.

* Remove unused imports

* Add back try/catch for the lint command

* Update test-fixture

* Revert and just hardcode the command name

* Revert change to vite.config.ts

* Add missing otel api dep

* Make use of fs-extra json funcs

* Replace require with import

* Simplify the exitCode setting

* Fix telemetry yargs option

* Introduce background job helper function

The helper ensure's we maintain the spawn options we need to support different platforms. It also ensures the output is written to a log file within the '.redwood' folder.

Reworks the new telemetry background process to use this helper. Introduces some relatively verbose output to the send process now that we do not need to be quiet.

* lint

* refactor telemetry enabling

---------

Co-authored-by: Dominic Saadi <dominiceliassaadi@gmail.com>