feat: Configure In-Memory or Redis-backed GraphQL Live Query and Subscriptions (#8791)
This PR refactors the Live Query and Subscription configuration for
Redwood Realtime to move more boilerplate setup to manage the stores and
behavior into the framework rather than in the app.

Redis or in-memory stores can be setup. One might use in memory for dev,
and then Redis for prod, etc.

See:
https://the-guild.dev/graphql/yoga-server/docs/features/subscriptions#distributed-pubsub-for-production

Redis and IORedis clients have been tested; Upstash Redis doesn't seem
to work as it can publish but not subscribe? Have to confirm.