chore: try running CI on pushes to next and release branches (#8844)
Main is becoming an increasingly different branch from the next and
release branches. Just because CI passes on main doesn't guarantee that
a change will work in the stable release. (More often than not it does,
but we need to start being more careful if we can.) I'd like to try
adding CI on pushes to next and release branches, so that we can get
feedback when commits are cherry picked, which is roughly once a day, or
once every other day.