Update intermission.md including a caution block to let the reader know not to look for a Typescript version (#8885)
I'm a first-time Redwood user. While working through the tutorial, once
I hit the intermission and started working through the examples, I
realized cloning the tutorial repo with the default instructions
contained js-only code. Since the tutorial provided example snippets in
Typescript, I figured there must be a separate branch or way to download
a typescript version. After poking around, I found this open issue
https://github.com/redwoodjs/redwood-tutorial/issues/56 and realized
this was in progress, so this just edits the documentation to save other
readers time while working through the tutorial. I'm no copywriter, and
the tutorial seems to take a specific tone, so please feel free to edit
the copy or make suggestions (i.e. link to the github issue in the
comment so people can track it) I didn't include the issue since there
doesn't seem to be a precedent for doing so in the rest of the
documentation. Using Redwood for the first time has been a great
experience by the way!

---------

Co-authored-by: Dominic Saadi <dominiceliassaadi@gmail.com>