fix(graphql-server): Allow correct context behaviour in testing (#8925)
**Problem**
The v6 RC has shown that during tests the context was not behaving
correctly after our modifications to request context isolation. As a
result of our changes the context object would always be an empty object
during testing.

This was only the case during testing. In standard operation the context
was isolated as expected.

**Changes**
This PR introduces three changes:
1. Fixes a slight error in the file ignore regex for the framework sync
tool. This was preventing the `testing` package from being synced.
2. This removes the recently added changes to the directive testing
helper function which used async local storage to ensure context was
behaving correctly during such testing.
3. Adds a `beforeEach` that is applied to every redwood project test
that mocks out the returned value of `getAsyncStoreInstance`. This
returned value is mocked to ensure that a new, empty, context is
received during each individual test.

**Outstanding**
1. Validate the context is still correctly isolated in production - just
as a super precaution.
2. Clear the spy I created in the `beforeEach` - be a good jest citizen?

**Notes**
1. I'm not sure of the performance implications that change will have on
massive test suites. I don't have such a test project to benchmark with.
I believe it is likely better to have tests with the correct behaviour
than faster tests.
2. I flip-flopped slightly on the mocked return value of
`getAsyncStoreInstance().getStore`. I think returning an object which is
lacking the full functionality will ensure that should we reach for this
behaviour in the future then we will be aware it may not work as
expected without further changes.
3. A smarter or more determined person might have been able to keep the
`getAsyncStoreInstance` function in the file it was original in. Right
now I could only get it to work by putting it in its own file.