fix(cli): Escape telemetry in exec command (#8947)
**Problem**
We now use OpenTelemetry to instrument the CLI. This means that all
telemetry generated during the CLI process is gathered and send off to
our backend for storage and later analysis. This helps us understand
usage and improve the CLI and overall framework.

We also allow users to instrument their project with OpenTelemetry for
their own usage. This becomes a problem when we execute user code within
the CLI process. This is what happens with `rw exec [script]` where we
directly call their script code within the same CLI process. We must
suppress telemetry gathering while executing their code or we risk
gathering _their_ telemetry as our own.

This does not apply when we spawn a new node process to run the user
code, such as `rw dev` or `rw serve`.

**Changes**
1. We now depend on `@opentelemetry/core` - I asked on their slack if I
could achieve the escaping without reaching into core and sticking with
api. I didn't get an answer but I can revisit this later if we need to
slim down the CLI again.
2. We have to include a small escape wrapper around the point where we
actually call user code within the exec command handler.

**Notes**
We never actually recorded any user project telemetry. This is proactive
rather than reactive.