@redwoodjs/web-server (#8916)
Add a new server for the Web side that docker files can use to serve the
web side of a RW project without depending on other big RW packages like
cli or internal.

The goal is that this should be a non-breaking change. If it's not, I
need to fix that.