fix(firebase): improve error handling (#8976)
Auditing firebase again for Redwood's v6 release, it was failing and I
forgot why. Referring back to
https://github.com/redwoodjs/redwood/pull/8305 reminded me that the
versions of `firebase-admin` between the framework and a user's project
need to be the same. One of the ways they get out of sync is with
renovate. I don't want to read from the firebase api package during the
firebase setup command—plus the versions can get out of sync in other
ways—so for now, may as well save everyone some time by making the error
more explicit:

<img width="2012" alt="image"
src="https://github.com/redwoodjs/redwood/assets/32992335/f1090965-5199-42d5-8831-3978e9a68050">

This coupling really isn't ideal at all since any time we bump the
version of `firebase-admin` in the framework and release a new versoin,
users have to bump it in their projects too. I'm not sure what a better
way is yet.

I've also upgraded the version of `firebase` (the package used on the
web side) that the setup command installs to v10 since it worked
locally.