fix: resolves Tailwind setup command failing due to unsupported prettier-plugin-tailwindcss v0.5 (#9076)
fixes #9075

`prettier-plugin-tailwindcss` >= 0.5 requires Prettier v3. This PR pins
the version to `0.4.1`, which is the most recent version prior to v0.5.

**NOTE:** This issue will still affect all previous versions of Redwood
because the setup ui tailwindcss command defaults to adding the Tailwind
deps at latest.

### Next Steps
I'll open two PRs, one to revert this and one to upgrade Prettier to v3
(both for next major).