feat(cli): lazy install data migrate and make it runnable on its own (#8572)
This PR makes the data migrate CLI command (`yarn rw data-migrate`) a
plugin like https://github.com/redwoodjs/redwood/pull/8454 did for
storybook. But this PR goes a little further, also adding a bin that
runs the data migrate's `up` command. I don't think it makes sense for
all of Redwood's CLI commands to do this, but it does for some, and data
migrate is one since it's an important part of Docker deploys.

- This PR extracts the Babel config out of `@redwoodjs/internal` into
its own package, `@redwoodjs/babel-config`. If I didn't do this,
`@redwoodjs/cli-data-migrate` would depend on `@redwoodjs/internal`
because it needs the `registerApiSideBabelHook` function, which would
make it's package size unacceptable
- Even without internal, the new package has an unfortunate dependency
on `@redwoodjs/structure` that I'm not going to try to refactor in this
PR
- Just noting that `@redwoodjs/structure` doesn't depend on any other
Redwood packages besides `@redwoodjs/project-config`, so I'm sure we
could make it smaller if we took the time to focus on it
- The Babel config in `@redwoodjs/internal` had a dependency on the
`typescript` package that I had to refactor as well
- The Babel config needs a lot of TLC, but strictly speaking it'd be
breaking so I'm holding off on that for now again. I'll open another PR
after this one that has those changes