Minor form improvements (#9050)
This PR addresses three minor items with the forms package:

1. If using on a field the `validation={{valueAsNumber: true}}` feature,
and `emptyAs` is set to `'undefined'`, the react-hook-form would
override the `emptyAs`. The solution is to remove the `valueAsNumber`
internally in this case and rely on the `setValueAs` coersion we have in
redwood.

2. If the `name` field is not set on a field component, the error would
be deep from within react-hook-forms and unintelligible. A change is
made to throw an intelligible error `"name field must be provided"`.
While typescript should normally help with this, not everyone uses
typescript, and I have been burned by this one even using typescript.

3. Fixed a typescript warning relating to a `ref` in the test file.

---------

Co-authored-by: Tobbe Lundberg <tobbe@tlundberg.com>