fix(babel-config): read TS config files using TS parser functions (#9121)
It looks like I have to add the `typescript` package, which is ~40 MB,
back to the `@redwoodjs/babel-config` package because TS config files
aren't valid JSON. I'm not sure what they are; this issue here
(https://github.com/microsoft/TypeScript/issues/44573) makes it sound
like they're their own format, which means we need to use TypeScript's
parser to be 100% sure we can read these files.

I had a look at how the `ts-morph` packages get around this problem.
They bundle using rollup. That's probably the answer for the future, but
we don't have experience bundling external dependencies like this, and
I'm trying not to introduce breaking changes, so leaving that for a
major and just taking the size hit for now.