chore(tasks): Benchmark tests (#8578)
Adds `yarn benchmark` which runs a series of k6 tests against the api
server (api serve).

This local tool should make it easier to catch failures or bad
performance which show only when under load. If you've refactored api
side code then you can run this task against main and then against your
branch to check for any obvious performance or correctness regressions.

This essentially does the following:
* Creates a new temp. redwood project like e2e testing
* Sets up a specific state of the project as defined by various setup
scripts
* Runs a series of k6 tests in those specifically setup environments

What this doesn't do:
* Handle testing different serving both dev, serve, etc.

Ideally in the future this can be adapted for use in CI too so that PR's
are automatically tested. This will also allow us to identify trends in
performance.