feat(cli): Setup command which uses an npm package (#8920)
**Problem**
We want to allow users to run arbitary setup commands that come from npm
packages. This is similar to the functionality of `npx` or `dlx` but we
want to allow additional redwood specific behaviour. Specifically for
now we want to ensure version compatibility between the setup package
and the redwood version of the user's project.

**Notes**
I published the package `@joshgmw/rw-setup-example` as a means of
testing the functionality in this PR.

**Changes**
1. Adds a `yarn rw setup package <npm-package>` command which will:
   1. Find the provided package on npm and check version compatibility
   2. Run `yarn dlx` with that package at the user project root

**Notes**
The UI of this isn't amazing but I don't really know what our consistent
design for the CLI is.

The naming of the command or where it sits in the command hierarchy is
something I'm happy to change if we feel the further need to.