fix(babel): use caret to specify some package versions (#9174)
the previous babel fix (https://github.com/redwoodjs/redwood/pull/9144)
to issues like https://github.com/redwoodjs/redwood/issues/9156 and
https://github.com/redwoodjs/redwood/issues/9073 was only a temporary
solution and didn't work for all users. after more thinking, i'm pretty
sure that the root of the problem is that we pin the version of the
babel packages we depend on.

take a redwood app that was just created from `yarn create redwood-app`.
most redwood projects request a specific version of a babel package. for
example, for `@babel/core`, redwood packages say they want `7.22.17`.
the redwood packages aren't the only ones that want `@babel/core`, but
they use a caret when they ask for the version, like `^7.22.5`

at the outset, yarn can usually make everyone happy with one version.
but say a new version of `@babel/core` comes out, `7.22.19`. and then a
user installs storybook, or some other package that depends on
`@babel/core`. the problem is that yarn checks npm to see if there's a
new version, and if there is, it updates the caret versions like
`^7.22.5` to resolve to the new one. but all the redwood packages are
insisting they want `7.22.17`, not `7.22.19`, so yarn has to keep both
around and now, when it comes to hoisting, has a choice to make: which
version of `@babel/core` is at `node_modules/@babel/core`? more packages
are ok with `7.22.19` so it wins out, and `7.22.17` gets buried into the
redwood package's node_modules.

to figure out which babel packages need the caret, i relied on a
combination of `yarn info -R --dependents ${packageName}`, `yarn why
${packageName}`, and `find ./node_modules -type d -path
'*/${packageName}'`.

the next steps for this are to get it into canary so i can confirm.